feat : 기존의 계획 수정 및 삭제 로직에 권한 검증 로직 추가

+++ src/main/java/com/example/Triple_clone/domain/plan/application/DetailPlanFacadeService.java
@@ -5,6 +5,8 @@ import com.example.Triple_clone.common.error.RestApiException;
 import com.example.Triple_clone.common.logging.logMessage.PlanLogMessage;
 import com.example.Triple_clone.domain.accommodation.application.AccommodationQueryService;
 import com.example.Triple_clone.domain.accommodation.domain.Accommodation;
+import com.example.Triple_clone.domain.member.application.UserService;
+import com.example.Triple_clone.domain.member.domain.Member;
 import com.example.Triple_clone.domain.plan.domain.DetailPlan;
 import com.example.Triple_clone.domain.plan.domain.Plan;
 import com.example.Triple_clone.domain.plan.web.dto.DetailPlanDto;
@@ -27,12 +29,20 @@ public class DetailPlanFacadeService {
     private final PlanService planService;
     private final DetailPlanService detailPlanService;
     private final RecommendService recommendService;
+    private final UserService userService;
     private final AccommodationQueryService accommodationService;
+    private final PlanPermissionUtils planPermissionUtils;
 
     @Transactional
-    public DetailPlanDto create(DetailPlanDto detailPlanDto) {
+    public DetailPlanDto create(DetailPlanDto detailPlanDto, String email) {
         Plan plan = planService.findById(detailPlanDto.planId());
         DetailPlan detailPlan = detailPlanDto.toEntity(plan);
+        Member member = userService.findByEmail(email);
+
+        if (!planPermissionUtils.hasEditPermission(plan, member)) {
+            log.warn(PlanLogMessage.PLAN_ACCESS_FAILED.format(email, plan));
+            throw new RestApiException(AuthErrorCode.AUTH_ERROR_CODE);
+        }
 
         detailPlanService.save(detailPlan);
         plan.addDetailPlan(detailPlan);
@@ -40,30 +50,42 @@ public class DetailPlanFacadeService {
     }
 
     @Transactional
-    public DetailPlanDto addRecommendation(long recommendationId, long planId) {
+    public DetailPlanDto addRecommendation(long recommendationId, long planId, String email) {
         Plan plan = planService.findById(planId);
         Recommendation recommendation = recommendService.findById(recommendationId);
 
         DetailPlanDto detailPlanDto = new DetailPlanDto(planId, recommendation.getLocation(), plan.getStartDay(), "00:00", null);
-        this.create(detailPlanDto);
+        this.create(detailPlanDto, email);
 
         return detailPlanDto;
     }
 
     @Transactional
-    public ReservationCreateDto createReservation(ReservationCreateDto reservationCreateDto) {
+    public ReservationCreateDto createReservation(ReservationCreateDto reservationCreateDto, String email) {
         Plan plan = planService.findById(reservationCreateDto.planId());
         Accommodation accommodation = accommodationService.findById(reservationCreateDto.accommodationId());
         DetailPlan detailPlan = reservationCreateDto.toEntity(plan, accommodation);
+        Member member = userService.findByEmail(email);
+
+        if (!planPermissionUtils.hasEditPermission(plan, member)) {
+            log.warn(PlanLogMessage.PLAN_ACCESS_FAILED.format(email, plan));
+            throw new RestApiException(AuthErrorCode.AUTH_ERROR_CODE);
+        }
 
         detailPlanService.save(detailPlan);
 
         return reservationCreateDto;
     }
 
-    public List<DetailPlanDto> readAll(long planId) {
+    public List<DetailPlanDto> readAll(long planId, String email) {
         Plan plan = planService.findById(planId);
+        Member member = userService.findByEmail(email);
         List<DetailPlanDto> response = new ArrayList<>();
+
+        if (!planPermissionUtils.hasViewPermission(plan, member)) {
+            log.warn(PlanLogMessage.PLAN_ACCESS_FAILED.format(email, plan));
+            throw new RestApiException(AuthErrorCode.AUTH_ERROR_CODE);
+        }
 
         for (DetailPlan detailPlan : plan.getPlans()) {
             response.add(detailPlan.toDto());
@@ -72,9 +94,15 @@ public class DetailPlanFacadeService {
     }
 
     @Transactional
-    public DetailPlanDto update(DetailPlanUpdateDto updateDto) {
+    public DetailPlanDto update(DetailPlanUpdateDto updateDto, String email) {
         Plan plan = planService.findById(updateDto.planId());
         DetailPlan detailPlan = detailPlanService.findById(updateDto.detailPlanId());
+        Member member = userService.findByEmail(email);
+
+        if (!planPermissionUtils.hasEditPermission(plan, member)) {
+            log.warn(PlanLogMessage.PLAN_ACCESS_FAILED.format(email, plan));
+            throw new RestApiException(AuthErrorCode.AUTH_ERROR_CODE);
+        }
 
         if (isContain(plan, detailPlan)) {
             DetailPlan updatedDetailPlan = detailPlanService.update(detailPlan, updateDto);
@@ -93,9 +121,15 @@ public class DetailPlanFacadeService {
     }
 
     @Transactional
-    public DetailPlan delete(long planId, long detailPlanId) {
+    public DetailPlan delete(long planId, long detailPlanId, String email) {
         Plan plan = planService.findById(planId);
         DetailPlan detailPlan = detailPlanService.findById(detailPlanId);
+        Member member = userService.findByEmail(email);
+
+        if (!planPermissionUtils.hasEditPermission(plan, member)) {
+            log.warn(PlanLogMessage.PLAN_ACCESS_FAILED.format(email, plan));
+            throw new RestApiException(AuthErrorCode.AUTH_ERROR_CODE);
+        }
 
         if (isContain(plan, detailPlan)) {
             detailPlanService.delete(detailPlan);
